requirements.txt
# Streamlit for UI
streamlit==1.29.0

# LangChain Core
langchain==0.2.12
langchain-community==0.2.11
langchain-groq==0.1.4
langchain-huggingface==0.0.3

# Vector stores
faiss-cpu==1.8.0

# Document loaders & parsing
pymupdf==1.23.8   # for PyMuPDFLoader

# Embeddings
sentence-transformers==2.7.0

# Environment variables
python-dotenv==1.0.1
